refactor - 2: move internal gitutil and builtins packages to pkg/lib/
