+++ commands/pkg/update/cmdupdate_test.go
@@ -26,10 +26,10 @@ import (
 
 	"github.com/kptdev/kpt/commands/pkg/get"
 	"github.com/kptdev/kpt/commands/pkg/update"
-	"github.com/kptdev/kpt/internal/gitutil"
 	"github.com/kptdev/kpt/internal/testutil"
 	"github.com/kptdev/kpt/internal/testutil/pkgbuilder"
 	kptfilev1 "github.com/kptdev/kpt/pkg/api/kptfile/v1"
+	"github.com/kptdev/kpt/pkg/lib/gitutil"
 	"github.com/kptdev/kpt/pkg/printer/fake"
 	"github.com/spf13/cobra"
 	"github.com/stretchr/testify/assert"
