+++ pkg/lib/builtins/testdata/pkg-with-existing-ctx/out.yaml
