+++ pkg/lib/builtins/testdata/pkg-with-nesting/in.yaml
