+++ pkg/lib/builtins/testdata/pkg-wo-nesting/in.yaml
