+++ pkg/lib/builtins/testdata/pkg-wo-nesting/out.yaml
